remove line  outVol.setObjLabel(vol.getObjLabel())
outVol.setObjLabel(vol.getObjLabel()) asigns to the new volume exactly the same label than the input volume has. In this way they are two different object with the same alias and therefore, there is no way to distinguish them. I did not remove the line:

            outVol.setObjComment(vol.getObjComment())

but I also think it should be gone