Use of dynamic dependencies (#905)
## Note: This PR needs to be merged after
[xmipp-995](https://github.com/I2PC/xmipp/pull/995).
- `xmipp3-installer` will now be automatically installed with the plugin
to ease Xmipp's new installation process.
- Instead of pointing to a specific Xmipp's version (`v3.24.0`, for
example), a new dynamic tag `v3` is used, to avoid needing to release a
new version of the plugin just because a new release of xmipp was
generated. Now releases in the plugin will only be tiggered by changes
in the code of the plugin itself, nothing else.

---------

Co-authored-by: Oier Lauzirika Zarrabeitia <oierlauzi@bizkaia.eu>
Co-authored-by: alberto <alber.mena@gmail.com>